adding tests migration for orders tests
