--- packages/augur-ui/test/orders/actions/update-orders-test.js
@@ -1,1 +1,0 @@
-describe("modules/orders/actions/update-orders.js", () => {});
