--- packages/augur-ui/test/orders/helpers/is-order-of-user-test.js
@@ -1,18 +1,0 @@
-describe("modules/orders/helpers/is-order-of-user.js", () => {
-  const {
-    isOrderOfUser
-  } = require("../../../src/modules/orders/helpers/is-order-of-user");
-
-  describe("isOrderOfUser", () => {
-    it("should return false if order is not of user", () => {
-      assert.isFalse(isOrderOfUser({ owner: "owner_address" }, null));
-      assert.isFalse(
-        isOrderOfUser({ owner: "owner_address" }, "some other address")
-      );
-    });
-
-    it("should return correct ", () => {
-      assert.isTrue(isOrderOfUser({ owner: "owner_address" }, "owner_address"));
-    });
-  });
-});
